// src/types/emotion.ts
export type Emotion =
  | 'neutral'
  | 'happy'
  | 'sad'
  | 'sleepy'
  | 'angry'
  | 'curious'
  | 'bored'
  | 'scared'
  | 'calm'
  | 'love'
  | 'excited'
  | 'confused'
  | 'surprised'
  | 'annoyed'
  | 'shy'
  | 'proud'
  | 'silly'
  | 'determined'
  | 'worried'
  | 'playful';

export const allEmotions: Emotion[] = [
  'neutral',
  'happy',
  'sad',
  'sleepy',
  'angry',
  'curious',
  'bored',
  'scared',
  'calm',
  'love',
  'excited',
  'confused',
  'surprised',
  'annoyed',
  'shy',
  'proud',
  'silly',
  'determined',
  'worried',
  'playful',
];